+++ greengift/src/test/java/com/apiserver/greengift/ApiServerApplicationTests.java
@@ -1,10 +1,10 @@
-package com.greengift.apiserver;
+package com.apiserver.greengift;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApiServerApplicationTests {
+class greengiftApplicationTests {
 
 	@Test
 	void contextLoads() {
